Handle EAGAIN when reading files
Apparently this can happen even when the file descriptor is in blocking mode,
as seen in this strace:

176595 read(0,  <unfinished ...>
176595 <... read resumed>0x7ffdf39ace94, 2092) = -1 EAGAIN (Resource temporarily unavailable)

The case at hand happens with https://code.uplex.de/misc/elbestream, which
has been written as intermediate relief for sound processing programs which
fail stream processing on some inputs.

When stdin of the program using libsndfile is handed over to vmsplice() here
https://code.uplex.de/misc/elbestream/blob/e42437c85432ef042743e3004e7c52295e3ae35f/elbestream.c#L168
the EAGAIN is seen repeatedly, though not always.